[kern] Move kern machine to hb-kern.hh
